requirements.txt
colorama==0.4.6
loguru==0.6.0
numexpr==2.8.4
numpy==1.21.6
packaging==22.0
pandas==1.3.5
pyqtgraph==0.12.4
PySide2==5.15.2.1
python-dateutil==2.8.2
pytz==2022.6
Semi-ATE-STDF==0.1.28
shiboken2==5.15.2.1
six==1.16.0
tables==3.7.0
win32-setctime==1.1.0
XlsxWriter==3.0.3
prettytable == 3.5.0
